Singleton.someProp with know==true now behaves the same whether it's called as a setter or not.
Since set is only called with non-nil non-empty settings, and args is empty in the non-setter hit to doesNotUnderstand, the effect is the same. Assigning via setter is semantically weird, since what takes place isn't an assignment, but it allows for ProxySpace like behavior which may be interesting. However, this DOES allow a clean, shorthand setup of singletons using a pattern like this:

Singleton.foo(\a, \b, \c);
Singleton.bar(1, 2, 2);